Moved internal/azureutil to storage service pack
