+++ service_packs/storage/azure/group/group.go
@@ -7,17 +7,17 @@ import (
 	"github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2018-02-01/resources"
 	"github.com/Azure/go-autorest/autorest/azure/auth"
 	"github.com/Azure/go-autorest/autorest/to"
-	azureutil "github.com/citihub/probr/service_packs/storage/azure"
+	"github.com/citihub/probr/service_packs/storage/azure"
 )
 
 // Create creates a new Resource Group in the default location (configured using the AZURE_LOCATION environment variable).
 func Create(ctx context.Context, name string) (resources.Group, error) {
-	log.Printf("[INFO] creating Resource Group '%s' in location: %v", name, azureutil.ResourceLocation())
+	log.Printf("[INFO] creating Resource Group '%s' in location: %v", name, azure.ResourceLocation())
 	return client().CreateOrUpdate(
 		ctx,
 		name,
 		resources.Group{
-			Location: to.StringPtr(azureutil.ResourceLocation()),
+			Location: to.StringPtr(azure.ResourceLocation()),
 		})
 }
 
@@ -29,12 +29,12 @@ func Get(ctx context.Context, name string) (resources.Group, error) {
 
 // CreateWithTags creates a new Resource Group in the default location (configured using the AZURE_LOCATION environment variable) and sets the supplied tags.
 func CreateWithTags(ctx context.Context, name string, tags map[string]*string) (resources.Group, error) {
-	log.Printf("[INFO] creating Resource Group '%s' on location: '%v'", name, azureutil.ResourceLocation())
+	log.Printf("[INFO] creating Resource Group '%s' on location: '%v'", name, azure.ResourceLocation())
 	return client().CreateOrUpdate(
 		ctx,
 		name,
 		resources.Group{
-			Location: to.StringPtr(azureutil.ResourceLocation()),
+			Location: to.StringPtr(azure.ResourceLocation()),
 			Tags:     tags,
 		})
 }
@@ -42,10 +42,10 @@ func CreateWithTags(ctx context.Context, name string, tags map[string]*string) (
 func client() resources.GroupsClient {
 
 	// Create an azure resource group client object via the connection config vars
-	c := resources.NewGroupsClient(azureutil.SubscriptionID())
+	c := resources.NewGroupsClient(azure.SubscriptionID())
 
 	// Create an authorization object via the connection config vars
-	authorizer := auth.NewClientCredentialsConfig(azureutil.ClientID(), azureutil.ClientSecret(), azureutil.TenantID())
+	authorizer := auth.NewClientCredentialsConfig(azure.ClientID(), azure.ClientSecret(), azure.TenantID())
 
 	authorizerToken, err := authorizer.Authorizer()
 	if err == nil {
